# asset-log/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "FieldError": {
        "properties": {
          "field": {
            "description": "エラーが発生した入力項目名",
            "example": "quantity",
            "type": "string"
          },
          "message": {
            "description": "項目ごとのエラー内容",
            "example": "数量は正の数である必要があります",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
...
      "ImportReport": {
        "description": "検証結果。dry-run のレスポンス、および本登録が失敗したときの422ボディ。",
        "properties": {
          "errors": {
            "description": "1件でもあれば本登録は全体が失敗する",
            "items": {
              "$ref": "#/components/schemas/ImportRowError"
            },
            "type": "array"
          },
          "to_insert": {
            "description": "挿入対象になる行数",
            "minimum": 0,
            "type": "integer"
          },
          "to_skip_duplicate": {
            "description": "重複としてスキップされる行数",
            "minimum": 0,
            "type": "integer"
          },
          "total_rows": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "total_rows",
          "to_insert",
          "to_skip_duplicate",
          "errors"
        ],
        "type": "object"
      },
      "ImportRequest": {
        "properties": {
          "csv_content": {
            "description": "CSV本文。ヘッダ行を含む",
            "example": "account_id,asset_id,kind,quantity,price,fee,traded_at\n...",
            "type": "string"
          }
        },
        "required": [
          "csv_content"
        ],
        "type": "object"
      },
      "ImportResult": {
        "description": "本登録が成功したときの結果。",
        "properties": {
          "inserted": {
            "minimum": 0,
            "type": "integer"
          },
          "skipped_duplicate": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "inserted",
          "skipped_duplicate"
        ],
        "type": "object"
      },
      "ImportRowError": {
        "properties": {
          "message": {
            "type": "string"
          },
          "row": {
            "description": "CSVの行番号（ヘッダ行を除く1始まり）",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "row",
          "message"
        ],
        "type": "object"
      },
...
      "PriceItem": {
        "properties": {
          "price": {
            "example": "2350.5",
            "type": "string"
          },
          "priced_on": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "priced_on",
          "price"
        ],
        "type": "object"
      },
      "PriceResponse": {
        "properties": {
          "price": {
            "type": "string"
          },
          "priced_on": {
            "format": "date",
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "priced_on",
          "price",
          "source",
          "updated_at"
        ],
        "type": "object"
      },
      "ProblemDetails": {
        "description": "RFC 9457 準拠のエラーレスポンス（`application/problem+json`）",
        "properties": {
          "detail": {
            "description": "人間向けの詳細メッセージ",
            "example": "口座が見つかりません",
            "type": "string"
          },
          "errors": {
            "description": "項目単位のエラー（422のみ。空の場合はフィールドごと省略される）",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "status": {
            "description": "HTTPステータスコード",
            "example": 404,
            "format": "int32",
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "description": "エラー種別の短い説明",
            "example": "Not Found",
            "type": "string"
          },
          "trace_id": {
            "description": "サーバーログと突き合わせるための識別子",
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "description": "エラー種別の安定した識別子。クライアントはこの値で分岐できる",
            "example": "/errors/not-found",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "trace_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "UpsertPricesRequest": {
        "properties": {
          "asset_id": {
            "format": "uuid",
            "type": "string"
          },
          "prices": {
            "description": "1リクエストあたり最大1000件",
            "items": {
              "$ref": "#/components/schemas/PriceItem"
            },
            "type": "array"
          },
          "source": {
            "description": "価格の出所。未指定なら \"manual\"",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "asset_id",
          "prices"
        ],
        "type": "object"
      },
      "UpsertPricesResponse": {
        "properties": {
          "upserted": {
            "description": "登録・更新された件数",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "upserted"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "jobToken": {
        "description": "バッチ実行用の固定トークン",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "NISA/iDeCo を含む資産・取引の管理API",
    "license": {
      "name": ""
    },
    "title": "asset-log API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/import/transactions": {
      "post": {
        "operationId": "import_transactions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            },
            "description": "全行の取込に成功した"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReport"
                }
              }
            },
            "description": "検証エラー。1件でもあれば何も挿入されない。ボディは ProblemDetails ではなく ImportReport"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "import"
        ]
      }
    },
    "/import/transactions/dry-run": {
      "post": {
        "operationId": "import_transactions_dry_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReport"
                }
              }
            },
            "description": "検証結果。DBには何も書き込まれない。errors が空でなければ本登録は失敗する"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "import"
        ]
      }
    },
...
    "/prices": {
      "post": {
        "operationId": "upsert_prices",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPricesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertPricesResponse"
                }
              }
            },
            "description": "登録・更新した件数"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "prices が空"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "銘柄が存在しない"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "件数超過、未来日、価格が負"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "assets"
        ]
      }
    },
    "/prices/{asset_id}": {
      "get": {
        "operationId": "get_price_history",
        "parameters": [
          {
            "description": "銘柄ID",
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "開始日（含む）",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "終了日（含む）",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PriceResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "価格履歴（日付の昇順）"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "銘柄が存在しない"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "assets"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "ローカル開発",
      "url": "http://localhost:8080"
    }
  ],
...
}